remove class evals
