+++ app/controllers/decidim/action_delegator/verifications/admin/application_controller.rb
@@ -2,7 +2,7 @@
 
 module Decidim
   module ActionDelegator
-    module Verification
+    module Verifications
       module Admin
         # This controller is the abstract class from which all other controllers of
         # this engine inherit.
