+++ app/forms/decidim/action_delegator/consultations/vote_form_override.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ActionDelegator
+    module Consultations
+      module VoteFormOverride
+        included do
+          attribute :decidim_consultations_delegation_id, Integer
+        end
+      end
+    end
+  end
+end
