+++ app/models/decidim/action_delegator/consultations/vote_override.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Decidim
+  module ActionDelegator
+    module Consultations
+      module VoteOverride
+        included do
+          has_paper_trail
+        end
+      end
+    end
+  end
+end
