--- app/overrides/forms/decidim/consultations/vote_form.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Decidim::Consultations::VoteForm.class_eval do
-  attribute :decidim_consultations_delegation_id, Integer
-end
