--- app/overrides/models/decidim/consultations/vote.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-Decidim::Consultations::Vote.class_eval do
-  has_paper_trail
-end
